perf(startup): keep Effect out of the CLI's eager import graph
`cli-app.ts` loads the agent stack only when a command actually runs, so
that `jazz --version` and `jazz --help` stay on the Commander tree. That
intent was being defeated one import deep: it reads two tier constants
from `types/peer`, which imports `types/disclosure-tier`, which imported
`Effect` and the tool-registry interface at module scope. Evaluating
Effect's module graph costs ~100-150ms, and every invocation paid it
before parsing argv.

Only `resolveToolAllowlist` needed either import, so it moves to its own
module; `disclosure-tier` now imports types alone, which the transpiler
elides. `allowedToolsForTier` and the tier constants stay where they are,
since they are what the doors and the CLI actually read.

Measured by `bench/startup.bench.ts`: 120.9ms -> 33.1ms for
`bun packages/runtime/src/main.ts --version`. Confirmed by import graph
rather than by wall clock alone — importing `cli-app` and then `effect`
went from 2.4ms for the second import (already loaded) to 101.6ms (no
longer loaded).

This is the source path, which is what contributors and `bun run cli`
pay. A released binary is compiled, so its startup is dominated by
loading the executable rather than by module evaluation, and is unchanged
by this.
