+++ packages/adapters/src/peers/serve.ts
@@ -33,8 +33,8 @@
 
 import { AgentRunner } from "@jazz/core/agent/agent-runner";
 import type { Agent } from "@jazz/core/types";
-import { resolveToolAllowlist } from "@jazz/core/types/disclosure-tier";
 import type { PeerConfig } from "@jazz/core/types/peer";
+import { resolveToolAllowlist } from "@jazz/core/types/resolve-tool-allowlist";
 import { generateConversationId } from "@jazz/core/utils/conversation-id";
 import { Effect } from "effect";
 import { record as recordLedger } from "./ledger";
